refactor: colocate FlowNodeInstancesTree
