+++ operate/client/src/App/ProcessInstance/FlowNodeInstanceLog/index.tsx
@@ -7,7 +7,7 @@
  */
 
 import React, {useRef} from 'react';
-import {FlowNodeInstancesTree} from '../FlowNodeInstancesTree';
+import {FlowNodeInstancesTree} from './FlowNodeInstancesTree';
 import {observer} from 'mobx-react';
 import {flowNodeInstanceStore} from 'modules/stores/flowNodeInstance';
 import {
